Remove observer queues from entity sets
With the addition of object pools that reuse entity references, any entity reference that is added to an entity set observer's queue isn't guaranteed to be valid the next time that queue is polled. To address this, I've simply removed the queues and implemented world events instead. The new world events use the world event bus and are processed immediately. The pre events can also be cancelled.